+++ ソースファイル/Character/Enemy/EnemyStateDash.cpp
@@ -23,7 +23,7 @@ namespace
 	//“®‚«‚Ì•ûŒü‚Ì”
 	constexpr int kMoveDirNum = 8;
 	//“®‚«‚Ì•ûŒü‚Ì”‚Ì”¼•ª
-	constexpr int kMoveDirNumHalf = kMoveDirNum * 0.5;
+	constexpr int kMoveDirNumHalf = static_cast<int>(kMoveDirNum * 0.5);
 	//ƒvƒŒƒCƒ„[‚É‹ß‚Ã‚«‚·‚¬‚È‚¢‚æ‚¤‚É
 	constexpr float kPlayerDistance = 15.0f;
 }
